requirements.txt
tensorflow==2.15.0
numpy==1.24.3
matplotlib==3.7.1
opencv-python==4.8.0.74
scikit-learn==1.3.0
pandas==1.5.3
seaborn==0.12.2
Pillow==9.5.0
keras==2.15.0
jupyter==1.0.0
ipykernel==6.24.0
streamlit